Refactor imagerunner
